run the isoextract executable on a vector of file paths this is usually not called directly

run the isoextract executable on a vector of file paths this is usually not called directly

Usage

ir_extract_isofiles(
  file_paths,
  pretty_json = FALSE,
  show_progress = is_interactive(),
  show_problems = TRUE
)

Arguments

file_paths

paths to the isodat file(s), single value or vector of paths. Use ir_find_isofiles() to get files in a folder.

pretty_json

whether to write the JSON output in human-readable pretty-printed format (default: FALSE). Useful for debugging; has no effect on the data read back by ir_read_isofiles(). Note that pretty-printed files are larger than compact ones.

show_progress

whether to show a progress bar, by default always enabled when running interactively e.g. inside Positron or RStudio (and disabled in a notebook), turn off with show_progress = FALSE

show_problems

whether to show problems encountered along the way (rather than just keeping track of them with ir_get_problems()). Set to show_problems = FALSE to turn off the live printout. Either way, all encountered problems can be retrieved with running ir_get_problems() for the returned list